move reducer to separate module
