+++ src/pages/CreatePost/CreatePostForm.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useReducer } from "react";
+import { useState, useEffect } from "react";
 import {
   Form,
   TextField,
@@ -12,138 +12,12 @@ import ColorSelector from "./ColorSelector";
 import { IBreed, IFetchedBreeds } from "../../services/pet";
 import ImageFileList from "./ImageFileList";
 import { getBreeds } from "../../services/pet";
+import useCreatePostForm, { ActionType } from "./useCreatePostForm";
 
 export interface ICreatePostFormProps {}
 
-interface ICreatePostFormState {
-  name: string;
-  species: string;
-  sex: string;
-  breedType: string;
-  breeds: IBreed[];
-  age: string;
-  weight: string;
-  microchip: string;
-  coatColors: string[];
-  eyeColors: string[];
-  imageFiles: File[];
-  description: string;
-}
-
-enum ActionType {
-  SET_NAME = "setName",
-  SET_SPECIES = "setSpecies",
-  SET_SEX = "setSex",
-  SET_BREEDS = "setBreeds",
-  SET_BREED = "setBreed",
-  ADD_BREED = "addBreed",
-  REMOVE_BREED = "removeBreed",
-  CLEAR_BREEDS = "clearBreeds",
-  SET_BREED_TYPE = "setBreedType",
-  SET_AGE = "setAge",
-  SET_WEIGHT = "setWeight",
-  SET_MICROCHIP = "setMicrochip",
-  SET_DESCRIPTION = "setDescription",
-  ADD_COAT_COLOR = "addCoatColor",
-  REMOVE_COAT_COLOR = "removeCoatColor",
-  CLEAR_COAT_COLORS = "clearCoatColors",
-  ADD_EYE_COLOR = "addEyeColor",
-  REMOVE_EYE_COLOR = "removeEyeColor",
-  CLEAR_EYE_COLORS = "clearEyeColors",
-  REMOVE_IMAGE_FILE = "removeImageFile",
-  ADD_IMAGE_FILE = "addImageFile",
-}
-
-interface IFormAction {
-  type: ActionType;
-  payload: any;
-}
-
-const initialState = {
-  name: "",
-  species: "",
-  sex: "",
-  breedType: "unknown",
-  breeds: [],
-  age: "",
-  weight: "",
-  microchip: "",
-  coatColors: ["red", "green", "blue"],
-  eyeColors: ["orange", "yellow", "purple"],
-  imageFiles: [],
-  description: "",
-};
-
-function reducer(
-  state: ICreatePostFormState,
-  action: IFormAction
-): ICreatePostFormState {
-  switch (action.type) {
-    case ActionType.SET_NAME:
-      return { ...state, name: action.payload };
-    case ActionType.SET_SPECIES:
-      return { ...state, species: action.payload };
-    case ActionType.SET_SEX:
-      return { ...state, sex: action.payload };
-    case ActionType.SET_BREEDS:
-      return { ...state, breeds: action.payload };
-    case ActionType.SET_BREED:
-      return { ...state, breeds: [action.payload] };
-    case ActionType.ADD_BREED:
-      return { ...state, breeds: [...state.breeds, action.payload] };
-    case ActionType.REMOVE_BREED:
-      return {
-        ...state,
-        breeds: state.breeds.filter((breed) => breed.id !== action.payload),
-      };
-    case ActionType.CLEAR_BREEDS:
-      return { ...state, breeds: [] };
-    case ActionType.SET_BREED_TYPE:
-      return { ...state, breedType: action.payload };
-    case ActionType.SET_AGE:
-      return { ...state, age: action.payload };
-    case ActionType.SET_WEIGHT:
-      return { ...state, weight: action.payload };
-    case ActionType.SET_MICROCHIP:
-      return { ...state, microchip: action.payload };
-    case ActionType.SET_DESCRIPTION:
-      return { ...state, description: action.payload };
-    case ActionType.ADD_COAT_COLOR:
-      return { ...state, coatColors: [...state.coatColors, action.payload] };
-    case ActionType.REMOVE_COAT_COLOR:
-      return {
-        ...state,
-        coatColors: state.coatColors.filter(
-          (value) => value !== action.payload
-        ),
-      };
-    case ActionType.CLEAR_COAT_COLORS:
-      return { ...state, coatColors: [] };
-    case ActionType.ADD_EYE_COLOR:
-      return { ...state, eyeColors: [...state.eyeColors, action.payload] };
-    case ActionType.REMOVE_EYE_COLOR:
-      return {
-        ...state,
-        eyeColors: state.eyeColors.filter((value) => value !== action.payload),
-      };
-    case ActionType.CLEAR_EYE_COLORS:
-      return { ...state, eyeColors: [] };
-    case ActionType.ADD_IMAGE_FILE:
-      return { ...state, imageFiles: [...state.imageFiles, action.payload] };
-    case ActionType.REMOVE_IMAGE_FILE:
-      return {
-        ...state,
-        imageFiles: state.imageFiles.filter((file, index) => {
-          return index !== action.payload;
-        }),
-      };
-    default:
-      return state;
-  }
-}
-
 export default function CreatePostForm(props: ICreatePostFormProps) {
-  const [state, dispatch] = useReducer(reducer, initialState);
+  const { state, dispatch, errors, errorDispatch } = useCreatePostForm();
   const [fetchedBreeds, setFetchedBreeds] = useState<IFetchedBreeds>({
     dogBreeds: [],
     catBreeds: [],
@@ -221,17 +95,19 @@ export default function CreatePostForm(props: ICreatePostFormProps) {
     dispatch({ type: ActionType.CLEAR_EYE_COLORS, payload: null });
   }
 
-  async function createPost() {
+  async function submit() {
     return;
   }
 
+  function validateFile(file: File) {
+    const maxSize = 1000 * 1024; // 1MB
+    if (file.size > maxSize) {
+      throw new Error("File too large. Maximum file size is 10MB.");
+    }
+  }
+
   return (
-    <Form
-      title="Create Post"
-      submitText="Create"
-      onSubmit={createPost}
-      error=""
-    >
+    <Form title="Create Post" submitText="Create" onSubmit={submit} error="">
       {/**
        * TODO:
        *
@@ -371,13 +247,22 @@ export default function CreatePostForm(props: ICreatePostFormProps) {
       <ImageFileField
         id="uploadPhoto"
         label="Upload Photo"
+        error={errors.fileUpload}
         onChange={(e) => {
           const files = e.target.files;
           if (!files) return;
           for (let i = 0; i < files.length; i++) {
+            const file = files.item(i) as File;
+            try {
+              validateFile(file);
+            } catch (error: any) {
+              errorDispatch({ type: "fileUpload", payload: error.message });
+              return;
+            }
+
             dispatch({
               type: ActionType.ADD_IMAGE_FILE,
-              payload: files.item(i),
+              payload: file,
             });
           }
         }}
